Add a ".resolve" alias for the resolve task (#582)
When automating interactions with the outputs of a k8s_object,
it's easy to find and operate on tasks like ".apply" because there
is a sigil that you can search for in the target output. Because there
is no corresponding sigil for the "resolve" target (just the bare
target name) it's difficult to filter for this task. Looking for the
absence of "apply", "delete", etc is not very future-proof, since other
target types could be introduced.